tests: move ln graph definitions and prepare_invoice to lnhelpers.py
This makes them reusable from test files other than test_lnpeer and
removes the deferred test_lnpeer import in prepare_chans_and_peers_in_graph.
